Phase 1: extract features/share-links.js
shareDiagramLink/checkForDiagramLink + private showShareToast move out, with the
share-only MAX_DIAGRAM_URL_PARAM_LENGTH const. Opening a shared link creates a
tab (tabs core), supplied via configureShareLinks; the DI var is named openTab
(not createTab) to avoid shadowing main.js's global createTab under the eval
test harness. main.js ~1,736 -> ~1,688. Build + lint clean; 297 tests pass.
